requirements.txt
binance-connector=3.9.0